Refactor CSharpWriter into partial classes
