+++ src/LingoEngine.Lingo.Core/CSharpWriter.cs
@@ -10,7 +10,7 @@ namespace LingoEngine.Lingo.Core;
 /// Writes a small subset of Lingo AST to C# using the visitor pattern.
 /// Only a handful of nodes are supported for now.
 /// </summary>
-public class CSharpWriter : ILingoAstVisitor
+public partial class CSharpWriter : ILingoAstVisitor
 {
     private readonly StringBuilder _sb = new();
     private readonly string _methodAccessModifier;
@@ -86,208 +86,9 @@ public class CSharpWriter : ILingoAstVisitor
     private void Indent() => _indent++;
     private void Unindent() { if (_indent > 0) _indent--; }
 
-    private static string DatumToCSharp(LingoDatum datum)
-    {
-        return datum.Type switch
-        {
-            LingoDatum.DatumType.Integer or LingoDatum.DatumType.Float => datum.AsString(),
-            LingoDatum.DatumType.String => $"\"{EscapeString(datum.AsString())}\"",
-            LingoDatum.DatumType.Symbol => $"Symbol(\"{datum.AsString()}\")",
-            LingoDatum.DatumType.VarRef => datum.AsString(),
-            LingoDatum.DatumType.List => datum.Value is List<LingoNode> list
-                ? "[" + string.Join(", ", list.Select(n => Write(n).Trim())) + "]"
-                : "[]",
-            LingoDatum.DatumType.ArgList or LingoDatum.DatumType.ArgListNoRet => datum.Value is List<LingoNode> argList
-                ? string.Join(", ", argList.Select(n => Write(n).Trim()))
-                : string.Empty,
-            LingoDatum.DatumType.PropList => datum.Value is List<LingoNode> propList
-                ? "new LingoPropertyList<object?>" +
-                    (propList.Count == 0
-                        ? "()"
-                        : " { " + string.Join(", ", Enumerable.Range(0, propList.Count / 2)
-                            .Select(i => $"[{Write(propList[2 * i]).Trim()}] = {Write(propList[2 * i + 1]).Trim()}")) + " }")
-                : "new LingoPropertyList<object?>()",
-            _ => datum.AsString()
-        };
-    }
-
-    private record PropDesc(string Name, string Comment, string DefaultValue);
-
-    private static string EscapeString(string s) => s
-        .Replace("\\", "\\\\")
-        .Replace("\"", "\\\"")
-        .Replace("\n", "\\n");
-
-    private static string FormatDefault(LingoDatum datum, string? format)
-    {
-        var value = datum.AsString();
-        if (string.Equals(format, "string", StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(format, "symbol", StringComparison.OrdinalIgnoreCase))
-        {
-            return $"\"{EscapeString(value)}\"";
-        }
-        return value;
-    }
-
-    private void WritePropertyDescriptionListHandler(LingoHandlerNode node)
-    {
-        Append(_methodAccessModifier);
-        Append(" BehaviorPropertyDescriptionList? GetPropertyDescriptionList()");
-        AppendLine();
-        AppendLine("{");
-        Indent();
-
-        var props = new List<PropDesc>();
-        foreach (var child in node.Block.Children)
-        {
-            if (child is LingoCallNode call &&
-                call.Callee is LingoVarNode v &&
-                v.VarName.Equals("addProp", StringComparison.OrdinalIgnoreCase) &&
-                call.Arguments is LingoDatumNode argDatum &&
-                argDatum.Datum.Type == LingoDatum.DatumType.ArgList &&
-                argDatum.Datum.Value is List<LingoNode> args &&
-                args.Count >= 3)
-            {
-                if (args[1] is not LingoDatumNode symNode) continue;
-                var propName = symNode.Datum.AsSymbol();
-                if (args[2] is not LingoDatumNode propList ||
-                    propList.Datum.Type != LingoDatum.DatumType.PropList ||
-                    propList.Datum.Value is not List<LingoNode> plist)
-                    continue;
-
-                string? comment = null;
-                string? format = null;
-                LingoDatum? defDatum = null;
-
-                for (int i = 0; i + 1 < plist.Count; i += 2)
-                {
-                    if (plist[i] is not LingoDatumNode keyNode) continue;
-                    var key = keyNode.Datum.AsSymbol();
-                    var valNode = plist[i + 1];
-                    if (key == "comment")
-                    {
-                        if (valNode is LingoDatumNode dn) comment = dn.Datum.AsString();
-                    }
-                    else if (key == "format")
-                    {
-                        if (valNode is LingoDatumNode dn) format = dn.Datum.AsSymbol();
-                    }
-                    else if (key == "default")
-                    {
-                        if (valNode is LingoDatumNode dn) defDatum = dn.Datum;
-                    }
-                }
-
-                if (propName != null && comment != null && defDatum != null)
-                {
-                    var defVal = FormatDefault(defDatum, format);
-                    props.Add(new PropDesc(propName, EscapeString(comment), defVal));
-                }
-            }
-        }
-
-        if (props.Count > 0)
-        {
-            AppendLine("return new BehaviorPropertyDescriptionList()");
-            Indent();
-            for (int i = 0; i < props.Count; i++)
-            {
-                var p = props[i];
-                var line = $".Add(this, x => x.{p.Name}, \"{p.Comment}\", {p.DefaultValue})";
-                if (i == props.Count - 1)
-                {
-                    line += ";";
-                }
-                AppendLine(line);
-            }
-            Unindent();
-        }
-        else
-        {
-            AppendLine("return new BehaviorPropertyDescriptionList();");
-        }
-
-        Unindent();
-        AppendLine("}");
-        AppendLine();
-    }
-
     private void Unsupported(LingoNode node)
     {
         // Unhandled nodes are ignored for now
-    }
-
-    public void Visit(LingoHandlerNode node)
-    {
-        var prevHandler = _currentHandlerName;
-        var name = node.Handler?.Name ?? string.Empty;
-        _currentHandlerName = name;
-
-        if (name.Equals("getPropertyDescriptionList", StringComparison.OrdinalIgnoreCase))
-        {
-            WritePropertyDescriptionListHandler(node);
-            _currentHandlerName = prevHandler;
-            return;
-        }
-
-        if (name.Length > 0)
-        {
-            var pascal = char.ToUpperInvariant(name[0]) + name[1..];
-            var lower = name.ToLowerInvariant();
-            string? paramDecl = lower switch
-            {
-                "blur" or "focus" => string.Empty,
-                "keydown" or "keyup" => "LingoKeyEvent key",
-                "mousedown" or "mouseup" or "mousemove" or "mousewheel" or
-                "mousewithin" or "mouseleave" or "mouseenter" or "mouseexit" => "LingoMouseEvent mouse",
-                _ => null
-            };
-
-            if (paramDecl != null)
-            {
-                Append(_methodAccessModifier);
-                Append(" void ");
-                Append(pascal);
-                Append("(");
-                Append(paramDecl);
-                AppendLine(")");
-                AppendLine("{");
-                Indent();
-                if (!string.IsNullOrEmpty(paramDecl))
-                {
-                    var paramVar = lower is "keydown" or "keyup" ? "key" : "mouse";
-                    foreach (var a in node.Handler.ArgumentNames.Where(a => !a.Equals("me", StringComparison.OrdinalIgnoreCase)))
-                        AppendLine($"var {a} = {paramVar};");
-                }
-                node.Block.Accept(this);
-                Unindent();
-                AppendLine("}");
-                AppendLine();
-            }
-            else
-            {
-                Append(_methodAccessModifier);
-                Append(" void ");
-                Append(pascal);
-                Append("(");
-                var args = node.Handler.ArgumentNames
-                    .Where(a => !a.Equals("me", StringComparison.OrdinalIgnoreCase))
-                    .Select(a => $"object {a}");
-                Append(string.Join(", ", args));
-                AppendLine(")");
-                AppendLine("{");
-                Indent();
-                node.Block.Accept(this);
-                Unindent();
-                AppendLine("}");
-                AppendLine();
-            }
-        }
-        else
-        {
-            node.Block.Accept(this);
-        }
-        _currentHandlerName = prevHandler;
     }
 
     public void Visit(LingoErrorNode node) => AppendLine("// error");
@@ -574,68 +375,6 @@ public class CSharpWriter : ILingoAstVisitor
         Append(")");
     }
 
-    public void Visit(LingoObjPropExprNode node)
-    {
-        if (node.Object is LingoMemberExprNode member &&
-            node.Property is LingoVarNode propVarText)
-        {
-            var propName = propVarText.VarName;
-            if (propName.Equals("text", StringComparison.OrdinalIgnoreCase) ||
-                propName.Equals("line", StringComparison.OrdinalIgnoreCase) ||
-                propName.Equals("word", StringComparison.OrdinalIgnoreCase) ||
-                propName.Equals("char", StringComparison.OrdinalIgnoreCase))
-            {
-                Append("GetMember<ILingoMemberTextBase>(");
-                member.Expr.Accept(this);
-                if (member.CastLib != null)
-                {
-                    Append(", ");
-                    member.CastLib.Accept(this);
-                }
-                Append(").");
-                Append(char.ToUpperInvariant(propName[0]) + propName[1..]);
-                return;
-            }
-        }
-
-        if (node.Object is LingoVarNode objVar &&
-            objVar.VarName.Equals("me", StringComparison.OrdinalIgnoreCase) &&
-            node.Property is LingoVarNode propVar)
-        {
-            Append(char.ToUpperInvariant(propVar.VarName[0]) + propVar.VarName[1..]);
-            return;
-        }
-
-        if (node.Object is LingoTheExprNode theNode)
-        {
-            node.Object.Accept(this);
-            Append(".");
-            if (node.Property is LingoVarNode propVar2)
-            {
-                var name = propVar2.VarName;
-                if (theNode.Prop.Equals("actorlist", StringComparison.OrdinalIgnoreCase) &&
-                    name.Equals("append", StringComparison.OrdinalIgnoreCase))
-                {
-                    Append("Add");
-                }
-                else
-                {
-                    Append(char.ToUpperInvariant(name[0]) + name[1..]);
-                }
-            }
-            else
-            {
-                node.Property.Accept(this);
-            }
-        }
-        else
-        {
-            node.Object.Accept(this);
-            Append(".");
-            node.Property.Accept(this);
-        }
-    }
-
     public void Visit(LingoPlayCmdStmtNode node)
     {
         Append("Play(");
@@ -643,35 +382,11 @@ public class CSharpWriter : ILingoAstVisitor
         AppendLine(");");
     }
 
-    public void Visit(LingoThePropExprNode node)
-    {
-        Append("TheProp(");
-        node.Property.Accept(this);
-        Append(")");
-    }
-
-    public void Visit(LingoMenuPropExprNode node)
-    {
-        Append("MenuProp(");
-        node.Menu.Accept(this);
-        Append(", ");
-        node.Property.Accept(this);
-        Append(")");
-    }
-
     public void Visit(LingoSoundCmdStmtNode node)
     {
         Append("Sound(");
         node.Command.Accept(this);
         AppendLine(");");
-    }
-
-    public void Visit(LingoSoundPropExprNode node)
-    {
-        Append("Sound(");
-        node.Sound.Accept(this);
-        Append(").");
-        node.Property.Accept(this);
     }
 
     public void Visit(LingoCursorStmtNode node)
@@ -771,17 +486,6 @@ public class CSharpWriter : ILingoAstVisitor
         Append("]");
     }
 
-    public void Visit(LingoSpritePropExprNode node)
-    {
-        Append("Sprite(");
-        node.Sprite.Accept(this);
-        Append(").");
-        if (node.Property is LingoVarNode propVar)
-            Append(char.ToUpperInvariant(propVar.VarName[0]) + propVar.VarName[1..]);
-        else
-            node.Property.Accept(this);
-    }
-
     public void Visit(LingoChunkDeleteStmtNode node)
     {
         Append("DeleteChunk(");
@@ -798,8 +502,6 @@ public class CSharpWriter : ILingoAstVisitor
 
     public void Visit(LingoGlobalDeclStmtNode node) { }
 
-    public void Visit(LingoPropertyDeclStmtNode node) { }
-
     public void Visit(LingoInstanceDeclStmtNode node) { }
 
     public void Visit(LingoRepeatWhileStmtNode node)
@@ -812,22 +514,6 @@ public class CSharpWriter : ILingoAstVisitor
         node.Body.Accept(this);
         Unindent();
         AppendLine("}");
-    }
-
-    public void Visit(LingoMenuItemPropExprNode node)
-    {
-        Append("menuItem(");
-        node.MenuItem.Accept(this);
-        Append(").");
-        node.Property.Accept(this);
-    }
-
-    public void Visit(LingoObjPropIndexExprNode node)
-    {
-        node.Object.Accept(this);
-        Append(".prop[");
-        node.PropertyIndex.Accept(this);
-        Append("]");
     }
 
     public void Visit(LingoRepeatWithInStmtNode node)
@@ -896,162 +582,6 @@ public class CSharpWriter : ILingoAstVisitor
         Append(")");
     }
 
-    public void Visit(LingoCallNode node)
-    {
-        if (node.Callee is LingoVarNode varNode && varNode.VarName.Equals("voidp", StringComparison.OrdinalIgnoreCase))
-        {
-            node.Arguments.Accept(this);
-            Append(" == null");
-            return;
-        }
-
-        if (!string.IsNullOrEmpty(node.Name))
-        {
-            if (!string.IsNullOrEmpty(node.TargetType))
-            {
-                var param = node.TargetType.ToLowerInvariant();
-                Append("CallMovieScript<");
-                Append(node.TargetType);
-                Append(">(");
-                Append($"{param} => {param}.");
-                Append(node.Name);
-                Append("());");
-                AppendLine();
-            }
-            else
-            {
-                AppendLine($"{node.Name}();");
-            }
-            return;
-        }
-
-        if (node.Callee is LingoVarNode func)
-        {
-            var name = func.VarName.ToLowerInvariant();
-            List<LingoNode>? args;
-            if (node.Arguments is LingoDatumNode dn &&
-                dn.Datum.Type == LingoDatum.DatumType.ArgList &&
-                dn.Datum.Value is List<LingoNode> list)
-            {
-                args = list;
-            }
-            else
-            {
-                args = new List<LingoNode> { node.Arguments };
-            }
-
-            switch (name)
-            {
-                case "add" when args.Count >= 2:
-                    args[0].Accept(this);
-                    Append(".Add(");
-                    args[1].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "addat" when args.Count >= 3:
-                    args[0].Accept(this);
-                    Append(".AddAt(");
-                    args[1].Accept(this);
-                    Append(", ");
-                    args[2].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "addprop" when args.Count >= 3:
-                    args[0].Accept(this);
-                    Append(".Add(");
-                    args[1].Accept(this);
-                    Append(", ");
-                    args[2].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "deleteat" when args.Count >= 2:
-                    args[0].Accept(this);
-                    Append(".DeleteAt(");
-                    args[1].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "getat" when args.Count >= 2:
-                    args[0].Accept(this);
-                    Append(".GetAt(");
-                    args[1].Accept(this);
-                    Append(")");
-                    return;
-                case "setat" when args.Count >= 3:
-                    args[0].Accept(this);
-                    Append(".SetAt(");
-                    args[1].Accept(this);
-                    Append(", ");
-                    args[2].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "count" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".Count");
-                    return;
-                case "deleteone" when args.Count >= 2:
-                    args[0].Accept(this);
-                    Append(".DeleteOne(");
-                    args[1].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "deleteall" when args.Count >= 1:
-                    args[0].Accept(this);
-                    AppendLine(".DeleteAll();");
-                    return;
-                case "duplicate" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".Duplicate()");
-                    return;
-                case "getone" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".GetOne()");
-                    return;
-                case "getlast" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".GetLast()");
-                    return;
-                case "getavalue" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".GetAValue()");
-                    return;
-                case "ilk" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".Ilk()");
-                    return;
-                case "listp" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".ListP()");
-                    return;
-                case "tolist" when args.Count >= 1:
-                    args[0].Accept(this);
-                    Append(".ToList()");
-                    return;
-                case "sort" when args.Count >= 1:
-                    args[0].Accept(this);
-                    AppendLine(".Sort();");
-                    return;
-                case "setprop" when args.Count >= 3:
-                    args[0].Accept(this);
-                    Append(".SetProp(");
-                    args[1].Accept(this);
-                    Append(", ");
-                    args[2].Accept(this);
-                    AppendLine(");");
-                    return;
-                case "deleteprop" when args.Count >= 2:
-                    args[0].Accept(this);
-                    Append(".DeleteProp(");
-                    args[1].Accept(this);
-                    AppendLine(");");
-                    return;
-            }
-        }
-
-        node.Callee.Accept(this);
-        Append("(");
-        node.Arguments.Accept(this);
-        AppendLine(");");
-    }
 
     public void Visit(LingoVarNode node)
     {
